Conflicts:
	assembler/src/cap/mosaic.hpp
	assembler/src/debruijn/detail_coverage.hpp
	assembler/src/debruijn/graph_simplification.hpp
	assembler/src/debruijn/simplification.cpp
	assembler/src/debruijn/stats/debruijn_stats.hpp
	assembler/src/include/de/insert_size_refiner.hpp
	assembler/src/test/cap/tests/synthetic/test_generator.py

+++ assembler/src/debruijn/indices/edge_info_updater.hpp
@@ -20,7 +20,7 @@ class EdgeInfoUpdater {
         index_.PutInIndex(kmer, e, 0);
         for (size_t i = index_.k(), n = nucls.size(); i < n; ++i) {
             kmer <<= nucls[i];
-            index_.PutInIndex(kmer, e, (int)i - (int)index_.k() + 1);
+            index_.PutInIndex(kmer, e, unsigned(i - index_.k()) + 1);
         }
     }
